resume_scanner/requirements.txt
# Core dependencies
PyPDF2>=3.0.0
python-docx>=0.8.11
numpy>=1.24.0
pandas>=2.0.0

# NLP and ML
scikit-learn>=1.3.0
nltk>=3.8.0

# Optional: Advanced NLP (comment out if not using)
# spacy>=3.5.0
# transformers>=4.30.0

# Database
# For PostgreSQL (optional, SQLite is default)
# psycopg2-binary>=2.9.6

# Progress bars
tqdm>=4.65.0

# Additional utilities
python-dateutil>=2.8.2